+++ bouncer/commands/simple_runtime_upgrade.ts
@@ -12,17 +12,14 @@
 
 import path from 'path';
 import { simpleRuntimeUpgrade } from '../shared/simple_runtime_upgrade';
-import { SwapContext, testAllSwaps } from '../shared/swapping';
 import { executeWithTimeout } from '../shared/utils';
-
-const swapContext = new SwapContext();
+import { testAllSwaps } from '../tests/all_swaps';
 
 async function main(): Promise<void> {
   await simpleRuntimeUpgrade(path.dirname(process.cwd()));
 
   if (process.argv[2] === '-test') {
-    await testAllSwaps(swapContext);
-    swapContext.print_report();
+    await testAllSwaps.run();
   }
 }
