+++ bouncer/full_bouncer.sh
@@ -1,16 +1,16 @@
 set -e
 echo "Running full bouncer 🧪"
 ./setup_for_test.sh
-./tests/gaslimit_ccm.ts
+./commands/execute_test.ts "Gas-Limit-Ccm-Swaps"
 ./tests/all_concurrent_tests.ts $1
-./tests/rotates_through_btc_swap.ts
-./tests/btc_utxo_consolidation.ts
-./tests/rotation_barrier.ts
-./tests/minimum_deposit.ts
+./commands/execute_test.ts "Rotates-Through-BTC-Swap"
+./commands/execute_test.ts "BTC-UTXO-Consolidation"
+./commands/execute_test.ts "Rotation-Barrier"
+./commands/execute_test.ts "Minimum-Deposit"
 
 if [[ $LOCALNET == false ]]; then
   echo "🤫 Skipping tests that require localnet"
 else
   echo "🚀 Running tests that require localnet"
-  ./tests/swap_after_temp_disconnecting_chains.ts
+  ./commands/execute_test.ts "Swap-After-Disconnection"
 fi
