+++ bouncer/test_commands/DCA_test.ts
@@ -1,0 +1,4 @@
+#!/usr/bin/env -S pnpm tsx
+import { testDCASwaps } from '../tests/DCA_test';
+
+await testDCASwaps.execute();
