+++ bouncer/test_commands/broker_fee_collection_test.ts
@@ -1,0 +1,4 @@
+#!/usr/bin/env -S pnpm tsx
+import { testBrokerFeeCollection } from '../tests/broker_fee_collection';
+
+await testBrokerFeeCollection.execute();
