+++ bouncer/test_commands/evm_deposits_test.ts
@@ -1,4 +1,4 @@
 #!/usr/bin/env -S pnpm tsx
-import { testEvmDeposits } from '../shared/evm_deposits';
+import { testEvmDeposits } from '../tests/evm_deposits';
 
 await testEvmDeposits.execute();
