+++ bouncer/test_commands/fill_or_kill.ts
@@ -1,0 +1,4 @@
+#!/usr/bin/env -S pnpm tsx
+import { testFillOrKill } from '../tests/fill_or_kill';
+
+await testFillOrKill.execute();
