+++ bouncer/test_commands/fund_redeem.ts
@@ -1,0 +1,4 @@
+#!/usr/bin/env -S pnpm tsx
+import { testFundRedeem } from '../tests/fund_redeem';
+
+await testFundRedeem.execute();
