+++ bouncer/test_commands/gaslimit_ccm.ts
@@ -1,0 +1,4 @@
+#!/usr/bin/env -S pnpm tsx
+import { testGasLimitCcmSwaps } from '../tests/gaslimit_ccm';
+
+await testGasLimitCcmSwaps.execute();
