+++ bouncer/test_commands/lp_api_test.ts
@@ -1,0 +1,4 @@
+#!/usr/bin/env -S pnpm tsx
+import { testLpApi } from '../tests/lp_api_test';
+
+await testLpApi.execute();
