+++ bouncer/test_commands/multiple_members_governance.ts
@@ -1,0 +1,4 @@
+#!/usr/bin/env -S pnpm tsx
+import { testMultipleMembersGovernance } from '../tests/multiple_members_governance';
+
+await testMultipleMembersGovernance.execute();
