+++ bouncer/test_commands/polkadot_runtime_update.ts
@@ -1,0 +1,5 @@
+#!/usr/bin/env -S pnpm tsx
+import { testPolkadotRuntimeUpdate } from '../tests/polkadot_runtime_update';
+
+// Note: This test only passes if there is more than one node in the network due to the polkadot runtime upgrade causing broadcast failures due to bad signatures.
+await testPolkadotRuntimeUpdate.execute();
