+++ bouncer/test_commands/rotates_vaults.sh
