+++ bouncer/test_commands/swap_less_than_ED.ts
@@ -1,0 +1,4 @@
+#!/usr/bin/env -S pnpm tsx
+import { swapLessThanED } from '../tests/swap_less_than_existential_deposit_dot';
+
+await swapLessThanED.execute();
