--- bouncer/tests/broker_fee_collection_test.ts
@@ -1,4 +1,0 @@
-#!/usr/bin/env -S pnpm tsx
-import { testBrokerFeeCollection } from '../shared/broker_fee_collection';
-
-await testBrokerFeeCollection.execute();
