--- bouncer/tests/swap_less_than_ED.ts
@@ -1,4 +1,0 @@
-#!/usr/bin/env -S pnpm tsx
-import { swapLessThanED } from '../shared/swap_less_than_existential_deposit_dot';
-
-await swapLessThanED.execute();
